+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - GCGI-55: Test Elba JSON output against the schema
+- GCGI-62: Add Oncogenic_Binary to Elba config schema
 
 ## v0.0.2: 2020-10-28
 ### Added